# supererp-openapi-formatted.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SuperERP",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/workers/": {
            "get": {
                "tags": [
                    "workers"
                ],
                "summary": "Read Workers",
                "description": "Retrieve workers.",
                "operationId": "workers-read_workers",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkersPublic"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "workers"
                ],
                "summary": "Create Worker",
                "description": "Create new worker.",
                "operationId": "workers-create_worker",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WorkerCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkerPublic"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/workers/{id}": {
            "get": {
                "tags": [
                    "workers"
                ],
                "summary": "Read Worker",
                "description": "Get worker by ID.",
                "operationId": "workers-read_worker",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkerPublic"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "workers"
                ],
                "summary": "Update Worker",
                "description": "Update a worker.",
                "operationId": "workers-update_worker",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WorkerUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkerPublic"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "workers"
                ],
                "summary": "Delete Worker",
                "description": "Delete a worker.",
                "operationId": "workers-delete_worker",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkerPublic"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
            "WorkerCreate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "title": "Name"
                    },
                    "gender": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 50
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Gender"
                    },
                    "department": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 255
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Department"
                    },
                    "address": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 500
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Address"
                    },
                    "aadhar": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 20
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Aadhar"
                    },
                    "bankname": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 255
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Bankname"
                    },
                    "ifscode": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 20
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Ifscode"
                    },
                    "accountno": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 30
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Accountno"
                    },
                    "pfno": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 30
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Pfno"
                    },
                    "esicno": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 30
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Esicno"
                    }
                },
                "type": "object",
                "required": [
                    "name"
                ],
                "title": "WorkerCreate"
            },
            "WorkerPublic": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "title": "Name"
                    },
                    "gender": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 50
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Gender"
                    },
                    "department": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 255
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Department"
                    },
                    "address": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 500
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Address"
                    },
                    "aadhar": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 20
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Aadhar"
                    },
                    "bankname": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 255
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Bankname"
                    },
                    "ifscode": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 20
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Ifscode"
                    },
                    "accountno": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 30
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Accountno"
                    },
                    "pfno": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 30
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Pfno"
                    },
                    "esicno": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 30
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Esicno"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "owner_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Owner Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "id",
                    "owner_id"
                ],
                "title": "WorkerPublic"
            },
            "WorkerUpdate": {
                "properties": {
                    "name": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 255
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Name"
                    },
                    "gender": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 50
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Gender"
                    },
                    "department": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 255
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Department"
                    },
                    "address": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 500
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Address"
                    },
                    "aadhar": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 20
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Aadhar"
                    },
                    "bankname": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 255
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Bankname"
                    },
                    "ifscode": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 20
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Ifscode"
                    },
                    "accountno": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 30
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Accountno"
                    },
                    "pfno": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 30
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Pfno"
                    },
                    "esicno": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 30
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Esicno"
                    }
                },
                "type": "object",
                "title": "WorkerUpdate"
            },
            "WorkersPublic": {
                "properties": {
                    "data": {
                        "items": {
                            "$ref": "#/components/schemas/WorkerPublic"
                        },
                        "type": "array",
                        "title": "Data"
                    },
                    "count": {
                        "type": "integer",
                        "title": "Count"
                    }
                },
                "type": "object",
                "required": [
                    "data",
                    "count"
                ],
                "title": "WorkersPublic"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/api/v1/login/access-token"
                    }
                }
            }
        }
    }
}